saasrest/requirements.txt
aiohttp==3.4.4
aioredis==1.2.0
amqp==2.4.1
appdirs==1.4.3
asgi-cors==0.3
asgiref==3.1.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.3.3
Automat==0.7.0
backports.csv==1.0.7
billiard==3.5.0.5
black==18.9b0
BlurWal==1.0.1
CacheControl==0.12.5
case==1.5.3
celery==4.2.1
certifi==2018.10.15
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
curlify==2.2.1
daphne==2.3.0
decorator==4.4.0
defusedxml==0.5.0
diff-match-patch==20181111
distlib==0.2.8
distro==1.4.0
Django==2.2.6
django-allauth==0.38.0
django-braces==1.13.0
django-colorfield==0.1.15
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-debug-toolbar-request-history==0.0.10
django-filter==2.1.0
django-import-export==1.2.0
django-mjml==0.7.0
django-money==0.14.4
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
django-rest-framework-social-oauth2==1.1.0
django-sslserver==0.20
djangorestframework==3.9.1
djangorestframework-api-key==1.4.1
dnspython==1.15.0
ecdsa==0.13.3
et-xmlfile==1.0.1
eventlet==0.24.1
ewmh==0.1.6
facebook-business==4.0.8
flake8==3.6.0
future==0.17.1
geocoder==1.38.1
greenlet==0.4.15
hiredis==1.0.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
jdcal==1.4.1
kombu==4.3.0
linecache2==1.0.0
lockfile==0.12.2
mccabe==0.6.1
mock==2.0.0
monotonic==1.5
msgpack==0.6.1
multidict==4.4.2
nose==1.3.7
numpy==1.15.4
oauth2-provider==0.0
oauthlib==2.1.0
odfpy==1.4.0
openpyxl==2.6.2
packaging==19.0
pbr==5.1.3
pep517==0.5.0
Pillow==6.2.0
pipenv==2018.11.26
progress==1.5
psycopg2-binary==2.8
py-moneyed==0.8.0
pyasn1==0.4.4
pycodestyle==2.4.0
pyflakes==2.0.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pynvim==0.3.2
pyparsing==2.3.1
python-engineio==3.10.0
python-jose==3.0.1
python-memcached==1.59
python-social-auth==0.3.6
python-socketio==2.0.0
python-xlib==0.25
python3-openid==3.1.0
pytoml==0.1.20
pytz==2019.1
PyYAML==5.1
ratelim==0.1.6
requests==2.22.0
requests-oauthlib==1.0.0
retrying==1.3.3
rsa==4.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
sqlparse==0.3.0
stripe==2.29.0
tablib==0.13.0
toml==0.10.0
traceback2==1.4.0
Twisted==19.7.0
txaio==18.8.1
unittest2==1.1.0
urllib3==1.25.6
vine==1.2.0
virtualenv==16.0.0
virtualenv-clone==0.5.3
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
yarl==1.2.6
zope.interface==4.6.0
